Add --force-ignore-with-dependencies option
Allow matching packages with regular expression which should be used
during to solution phase, but not install during the install phase.

Matched packages will be added to the dependency tree, but all their
requirements are wiped. Finally they are excluded from the list of
packages to install too.

Indirecte dependencies of an ignored package can still be pulled in by
dependencies of other packages.

Signed-off-by: Roman Mohr <rmohr@redhat.com>