Rename -f/--fedora-base-system to --basesystem
We are aiminig to support more than Fedora now, and the command
line options should match this intent.

Dropping the dash from the option name make it consistent with
existing options such as --nobest and --repofile. No short option
is allocated, because all the sensible choices are already taken.

The old options still work, but they are marked as deprecated.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>