Deprecate more shorthands
These shorthands are not ambiguous, but assigning shorthands to
options that are only used for a single command means they're no
longer available for options that are used across multiple
commands.

The shorthand "s", in particular, could be a reasonable candidate
for "basesystem", but we can't use it without introducing
ambiguity as long as rpm2tar's "symlinks" option also uses it.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>