Fix usage of bash arrays
After the recent changes to the way bash arrays are used, if
you have a rule like

  bazeldnf(
    name = "ldd_test",
    command = "ldd",
    ...
  )

then calling

  $ bazel run //:ldd_test

will result in

  Error: unknown command "\"ldd" for "bazeldnf"

This is caused by the excessive number of (nested, escaped)
quotes that are present in the expression. The new version
follows the recommendations from

  https://stackoverflow.com/questions/7577052/

accurately and as such doesn't exhibit the problem.

Fixes: 80ecb748102eb1a3da7ffd7a20e75c35757dd7b7
Signed-off-by: Andrea Bolognani <abologna@redhat.com>