set build mode to external for all deps
Consult the following resources for more information:

- https://github.com/bazelbuild/bazel-gazelle/issues/1217
- https://github.com/bazelbuild/bazel-gazelle/pull/1264
- https://github.com/bazelbuild/bazel-gazelle/pull/1201

TL;DR: When using recent versions of gazelle, the newer dependency resolution mode "static" is used by default. This especially fails for golang.org/x/... and when using a macro (_maybe) for the go_repository rules [1]. To avoid the issue, set build mode to external for now.

[1] https://github.com/bazelbuild/bazel-gazelle/issues/1217#issuecomment-1220987206