Support file:// scheme for repo URLs
In a repo.yaml file, instead of just the http:// or https:// scheme,
the user might want to specify a local path for an repo, e.g.

    repositories:
    - arch: x86_64
      baseurl: file:///RPMS
      name: local

And then use that with fetch, e.g.

    % bazeldnf fetch
    INFO[0000] Resolving repomd.xml from file:///RPMS/repodata/repomd.xml
    INFO[0000] Loading primary file from file:///RPMS/repodata/<HASH>-primary.xml.gz

To address this, update the Getter implementation to parse the URL
using net/url Parse, and the call a handler that opens a local file
and returns a wrapper http.Response when the scheme is "file".