Reconfigure the RPM URLs if checksum changed.
Currently if for some reason the RPM entry changed,
bazeldnf won't reconfigure the URLs if they are already
set. While this behavior aims to cover cases like mirror
urls getting added out of band, there is a case where the
URLs are messed up and do match the checksum. Therefore
with this change it reconfigures the RPM URLs when the
checksum changed.

Signed-off-by: Yufeng Duan <55268016+pomodorox@users.noreply.github.com>
Signed-off-by: Yufeng Duan <55268016+didovesei@users.noreply.github.com>