Fix name collision with tar2files command
If two files have the same name but are found in different
subdirectories then tar2files will fail.  An example of this is found
in abseil-cpp where there are multiple headers with the same name, but
sometimes in different subdirectories of `/usr/include`.

This change tweaks our tarfile handling so as to key our lookup map
based on the path minus anything before the expected prefix (this will
remove the `bazel-out/...` bits from the paths generated by the
tar2files rule.  This provides enough context to avoid the name
collisions but is still able to deal with the correct path lookup.