Add lockfile command to bazeldnf
In order to allow for easy generation of the lockfile to use with the
rpmtree tags in the bazel module extension we need to add an
additional command for this purpose.

The `lockfile` subcommand is derived from a combination  the `rpmtree`
subcommand in conjunction with some work from Manuel Naranjo on a
`bzlmod` subcommand.