rpm: Fix up tar_test.go
The current version of the Tar2Files test depends on the fact that we
previously allowed you to just match the basenames and dump the files
in the root directory.  When consumed by the bazel rule for tar2files,
the files are correctly placed in the desired directory.

This change re-works the existing test as follows:

- provides a recursive directory structure as the expected result

- separates files from expected results

- qualifies the file by full path similarly to what the bazel rules do

- generate the nested directories which bazel rules do implicitly