Add null repository rule for missing RPMs
Depending on the consumption model, users may need to have a stable
set of targets and repository rules to depend on even if the
underlying RPMs are missing.  This change is the other side of
allowing some of our RPM targets to be missing: if they are missing we
can generate dummy repository rules so that we get stable deps.