SAT solver: Relax unique package provider clause for a package requires
There are cases where a resource can be provided by multiple rpms.  For
example, “gcc” can be provided by packages “gcc-11”(an empty rpm) as well as “gcc11”.
And “gcc-11” requires “gcc11”. So, if a package depends on “gcc-11”, then we need to
install both “gcc-11” and “gcc11”. The current bf.Unique clause on the package requires
 expects only one of them to be there in the final solution. So, both the packages are
needed, then the solution fails.
This change relax the bf.Unique cluase to bf.Or to make both the packages available
in the final resolution if needed.