Initial preparation for separation of Loader and Resolver
Resolver really consists of 3 steps (one of which is somewhat
vestigial):

* Load
* Construct
* Solve

This change lays the groundwork for rolling Load and Construct into a
separate type named "Loader" which will enable us to test the loading
and solving steps more discretely.  We do this by first creating a
"Loader" type and retaining most of the state variables in the
"Resolver" type while the "Loader" refers back to it.