Factor our SAT data model into separate struct
This change is the next step in separating the "Loader" and "Resolver"
types from each other.  The mechanism for communicating between the
two types ends up in a separate "Model" struct which is an output from
the Loader and an input to the Resolver.  This captures all of the
shared data between the two types and provides a read-only interface
for the Resolver to consume the data through.