Final separation of Loader and Resolver
This change is the final step in separating the Loader and Resolver
types from each other.

We do this by eliminating the final helper methods in Resolver and
just surfacing the Loader as its own fully formed entity.  The Loader
takes ownership of the model which is subsequently passed through to
the Resolver.