rpm: support hardlinks in tar2files
Some RPMs use hardlinks and so the associated tar file is also using
hardlinks. Support is added in PrefixFilter for the same. It is
already supported in Untar.

Since the link target should appear before the reference in the
archive, I used a second mapping "allMappedFiles" to find the output
path for the link since the entry will have already been removed from
"fileMap".