rpm: minor factoring and tracing for tar2files
I've added the tracing that I'd have found helpful when trying to
understand the earlier failures associated with hardlinks in
tar2files.

I've also changed the 'if/else if/else' block to a switch statement
since each branch is predicated on the same `entry.Typeflag`, and then
moved out the fileMap delete since it's always done, and must always
be done, after processing an entry.