rpm: use symlinks instead of hard links to represent links
When finding a link in the tar file, instead of preserving it as a
link (using os.Link), simply handle it in the same was as we handle
symlinks, that is, creating an output symlink. This was being
discussed on #150 and we determined that symlinks provide better
reliability with bazel.

At the very least, hard links are not properly represented in the
remote cache / RBE and end up being copies anyway, so it's better to
avoid the inconsistency.