rpm: convert links to relative symlinks
When we're extracting tar files built from RPMs, most symlink targets
are relative to the symlink itself, e.g.

    /usr/lib64/libfoo.so -> /usr/lib64/libfoo.so.1

In the case of (hard) links stored in the RPMs, they're relative to
the package root and so we can't just reuse the TypeSymlink code to
handle TypeLink in PrefixFilter. Instead, we find the relative path to
the link target and create that symlink instead in place of using a
(hard) link, e.g.

    /usr/include/com_err.h -> et/com_err.h

In this case, the archive contained:

   ./usr/include/com_err.h    LINK ./usr/include/et/com_err.h
   ./usr/include/et/com_err.h FILE -