rpm: support hardlinks in tar2files (#150)
* rpm: support hardlinks in tar2files

Some RPMs use hardlinks and so the associated tar file is also using
hardlinks. Support is added in PrefixFilter for the same. It is
already supported in Untar.

Since the link target should appear before the reference in the
archive, I used a second mapping "allMappedFiles" to find the output
path for the link since the entry will have already been removed from
"fileMap".

* rpm: minor factoring and tracing for tar2files

I've added the tracing that I'd have found helpful when trying to
understand the earlier failures associated with hardlinks in
tar2files.

I've also changed the 'if/else if/else' block to a switch statement
since each branch is predicated on the same `entry.Typeflag`, and then
moved out the fileMap delete since it's always done, and must always
be done, after processing an entry.

* rpm: use symlinks instead of hard links to represent links

When finding a link in the tar file, instead of preserving it as a
link (using os.Link), simply handle it in the same was as we handle
symlinks, that is, creating an output symlink. This was being
discussed on #150 and we determined that symlinks provide better
reliability with bazel.

At the very least, hard links are not properly represented in the
remote cache / RBE and end up being copies anyway, so it's better to
avoid the inconsistency.

* rpm: convert links to relative symlinks

When we're extracting tar files built from RPMs, most symlink targets
are relative to the symlink itself, e.g.

    /usr/lib64/libfoo.so -> /usr/lib64/libfoo.so.1

In the case of (hard) links stored in the RPMs, they're relative to
the package root and so we can't just reuse the TypeSymlink code to
handle TypeLink in PrefixFilter. Instead, we find the relative path to
the link target and create that symlink instead in place of using a
(hard) link, e.g.

    /usr/include/com_err.h -> et/com_err.h

In this case, the archive contained:

   ./usr/include/com_err.h    LINK ./usr/include/et/com_err.h
   ./usr/include/et/com_err.h FILE -