Architecture preferences awareness
This implements mult-arch support for dependency resolver (`sat`) and in some other places.

It is based on the following assumptions:
1. User can now match packages to install also by architecture:
   - allow/ignore regexes now match against `.<arch>`-appended string;
   - top-level package requests can also refer to specific package architecture or architecture+version combo.
2. There is now total ordering of architectures.
   We start with {arch, "noarch"}, with `arch` having higher priority.
3. `--nobest` disabled keeps best packages per name.
   Suboptimal architectures, versions or packages from less prioritized repos are discarded.
4. Architecture preference is analysed first. It's before repo priority and version of the package.
5. Consistently, the top-level required package is always selected from the best available architecture, even if there might exist a package for different arch in more prioritized repo.

Some tests for multi-arch resolution were added.
These tests are based on a rather fictionalized scenario of a package available for both x86_64 and `noarch` architectures.
But since `noarch` is not specifically treated in this code, these can also be considered tests of a solution for multiple architectures.

Under the assumption that we never encounter a package distributed for `noarch` and some other architecture, the default setup of single architecture (+`noarch`) should not be affected by this change.