Fix `explodePackageRequires` to correctly handle case when requirement can't be satisfied
The previous implementation in this case returned `bf.Not(bfunique)`,
which was reasonable in the first iteration of the loop, creating
var → ¬var
to prevent that variable (package) from being chosen.

Unfortunately, `bfunique` could accumulate some other formulas and past the first iteration this is no longer correct.

In addition, the logic was split into two-phase pass,
to report all missing dependencies, not only the first one.