Add option to ignore missing lockfile (and generate null RPM repositories)
Some use cases may require running queries over the build graph prior to having the correct lockfile.
Such queries would fail if they refer to targets/aliases yet to be created.

This change basically extends the logic of null RPM repositories for unresolved packages (in presence of lockfile) to all packages (in absence of lockfile).
As it silences the message about the need to update lock file, it's not the default behaviour and is enabled with `ignore_missing_lockfile` attr of the `config` tag.