Enforce single version constraint per package
So far the only constraint for not installing multiple packages of the same name came from `explodePackageRequires` that for each requirement required that at most one package version was satisfying it (`bf.Unique`).
However, this relied on packages being requested explicitly by name. If not, there was no other constraint limiting that.

Since, later in the code, resolved packages are referenced solely by name, selecting multiple versions resulted in alternate versions being overwritten or otherwise ignored.

In the future, this constraint may be relaxed to allow multiple variants of the same package if they differ by architecture.