Upgrade various bazel modules (#163)
* Upgrade rules_go dep in go.mod

rules_go in go.mod wasn't matching our actual bazel_dep which was
causing some spurious error messages.

* Upgrade various bazel modules

This change pulls all of our various bazel module deps to their (as of
now) head revisions:

* bazel_skylib = 1.8.0
* platforms = 1.0.0
* bazel_features = 1.38.0
* gazelle = 0.47.0
* protobuf = 33.1
* toolchains_protoc = 0.6.0
* bazel_skylib_gazelle_plugin = 1.8.2
* aspect_bazel_lib = 2.21.2
* rules_java = 9.0.3
* rules_cc = 0.2.14
* rules_pkg = 1.1.0

* Upgrade buildtools in go.mod

After upgrading the various packages, buildtools is now out of sync so
we need to upgrade it.

* Import rules_shell explicitly

sh_binary needs to come from rules_shell now so we need to import it
explicitly.

* Update generated bazelrc files

Our changes appear to break the bazelrc files so we need to update
this via:

 bazel run @@//.aspect/bazelrc:update_aspect_bazelrc_presets